requirements.txt
cycler==0.9.0
ipaddress==1.0.15
matplotlib==1.5.0
numpy==1.10.1
Pillow==3.0.0
pip==7.1.2
pycparser==2.14
pyparsing==2.0.6
python-dateutil==2.4.2
pytz==2015.7
queuelib==1.4.2
requests==2.8.1
six==1.10.0
wheel==0.26.0
scipy==0.7.0
